+++ VelcroPhysics.Unity/Tools/Triangulation/Delaunay/TriangulationUtil.cs
@@ -37,7 +37,7 @@ namespace VelcroPhysics.Tools.Triangulation.Delaunay
 
     internal class TriangulationUtil
     {
-        public static double EPSILON = 1e-12;
+        public static float EPSILON = 1e-12f;
 
         /// <summary>
         /// Requirements:
@@ -66,40 +66,40 @@ namespace VelcroPhysics.Tools.Triangulation.Delaunay
         public static bool SmartIncircle(TriangulationPoint pa, TriangulationPoint pb, TriangulationPoint pc,
                                          TriangulationPoint pd)
         {
-            double pdx = pd.X;
-            double pdy = pd.Y;
-            double adx = pa.X - pdx;
-            double ady = pa.Y - pdy;
-            double bdx = pb.X - pdx;
-            double bdy = pb.Y - pdy;
+            float pdx = pd.X;
+            float pdy = pd.Y;
+            float adx = pa.X - pdx;
+            float ady = pa.Y - pdy;
+            float bdx = pb.X - pdx;
+            float bdy = pb.Y - pdy;
 
-            double adxbdy = adx * bdy;
-            double bdxady = bdx * ady;
-            double oabd = adxbdy - bdxady;
+            float adxbdy = adx * bdy;
+            float bdxady = bdx * ady;
+            float oabd = adxbdy - bdxady;
 
             //        oabd = orient2d(pa,pb,pd);
             if (oabd <= 0)
                 return false;
 
-            double cdx = pc.X - pdx;
-            double cdy = pc.Y - pdy;
+            float cdx = pc.X - pdx;
+            float cdy = pc.Y - pdy;
 
-            double cdxady = cdx * ady;
-            double adxcdy = adx * cdy;
-            double ocad = cdxady - adxcdy;
+            float cdxady = cdx * ady;
+            float adxcdy = adx * cdy;
+            float ocad = cdxady - adxcdy;
 
             //      ocad = orient2d(pc,pa,pd);
             if (ocad <= 0)
                 return false;
 
-            double bdxcdy = bdx * cdy;
-            double cdxbdy = cdx * bdy;
+            float bdxcdy = bdx * cdy;
+            float cdxbdy = cdx * bdy;
 
-            double alift = adx * adx + ady * ady;
-            double blift = bdx * bdx + bdy * bdy;
-            double clift = cdx * cdx + cdy * cdy;
+            float alift = adx * adx + ady * ady;
+            float blift = bdx * bdx + bdy * bdy;
+            float clift = cdx * cdx + cdy * cdy;
 
-            double det = alift * (bdxcdy - cdxbdy) + blift * ocad + clift * oabd;
+            float det = alift * (bdxcdy - cdxbdy) + blift * ocad + clift * oabd;
 
             return det > 0;
         }
@@ -107,28 +107,28 @@ namespace VelcroPhysics.Tools.Triangulation.Delaunay
         public static bool InScanArea(TriangulationPoint pa, TriangulationPoint pb, TriangulationPoint pc,
                                       TriangulationPoint pd)
         {
-            double pdx = pd.X;
-            double pdy = pd.Y;
-            double adx = pa.X - pdx;
-            double ady = pa.Y - pdy;
-            double bdx = pb.X - pdx;
-            double bdy = pb.Y - pdy;
+            float pdx = pd.X;
+            float pdy = pd.Y;
+            float adx = pa.X - pdx;
+            float ady = pa.Y - pdy;
+            float bdx = pb.X - pdx;
+            float bdy = pb.Y - pdy;
 
-            double adxbdy = adx*bdy;
-            double bdxady = bdx*ady;
-            double oabd = adxbdy - bdxady;
+            float adxbdy = adx*bdy;
+            float bdxady = bdx*ady;
+            float oabd = adxbdy - bdxady;
             //        oabd = orient2d(pa,pb,pd);
             if (oabd <= 0)
             {
                 return false;
             }
 
-            double cdx = pc.X - pdx;
-            double cdy = pc.Y - pdy;
+            float cdx = pc.X - pdx;
+            float cdy = pc.Y - pdy;
 
-            double cdxady = cdx*ady;
-            double adxcdy = adx*cdy;
-            double ocad = cdxady - adxcdy;
+            float cdxady = cdx*ady;
+            float adxcdy = adx*cdy;
+            float ocad = cdxady - adxcdy;
             //      ocad = orient2d(pc,pa,pd);
             if (ocad <= 0)
             {
@@ -140,13 +140,13 @@ namespace VelcroPhysics.Tools.Triangulation.Delaunay
 
         public static bool InScanArea(TriangulationPoint pa, TriangulationPoint pb, TriangulationPoint pc, TriangulationPoint pd)
         {
-            double oadb = (pa.X - pb.X) * (pd.Y - pb.Y) - (pd.X - pb.X) * (pa.Y - pb.Y);
+            float oadb = (pa.X - pb.X) * (pd.Y - pb.Y) - (pd.X - pb.X) * (pa.Y - pb.Y);
             if (oadb >= -EPSILON)
             {
                 return false;
             }
 
-            double oadc = (pa.X - pc.X) * (pd.Y - pc.Y) - (pd.X - pc.X) * (pa.Y - pc.Y);
+            float oadc = (pa.X - pc.X) * (pd.Y - pc.Y) - (pd.X - pc.X) * (pa.Y - pc.Y);
             if (oadc <= EPSILON)
             {
                 return false;
@@ -162,9 +162,9 @@ namespace VelcroPhysics.Tools.Triangulation.Delaunay
         /// =  (x1-x3)*(y2-y3) - (y1-y3)*(x2-x3)
         public static Orientation Orient2d(TriangulationPoint pa, TriangulationPoint pb, TriangulationPoint pc)
         {
-            double detleft = (pa.X - pc.X) * (pb.Y - pc.Y);
-            double detright = (pa.Y - pc.Y) * (pb.X - pc.X);
-            double val = detleft - detright;
+            float detleft = (pa.X - pc.X) * (pb.Y - pc.Y);
+            float detright = (pa.Y - pc.Y) * (pb.X - pc.X);
+            float val = detleft - detright;
             if (val > -EPSILON && val < EPSILON)
             {
                 return Orientation.Collinear;
